Сделал 4 домашку и тренировочные задания
